refactor(bruno-js): share local module path resolver
